fix: Fix review comments
