+++ microceph/ceph/cephfs_snapshot.go
@@ -11,12 +11,14 @@ import (
 
 // Contains methods for interacting with the Ceph FS mirror snapshot funtionality
 
+// CephFSSnapshotMirrorPeer encapsulates information about a single CephFS snapshot mirror peer
 type CephFSSnapshotMirrorPeer struct {
 	ClientName string `json:"client_name"`
 	SiteName   string `json:"site_name"`
 	Volume     string `json:"fs_name"`
 }
 
+// cephFSSnapshotMirrorEnableVolume enables mirroring for a specific CephFS volume
 func cephFSSnapshotMirrorEnableVolume(volume string) error {
 	_, err := cephRun(cephFSSnapshotMirrorCmd([]string{
 		"enable", volume,
@@ -25,6 +27,7 @@ func cephFSSnapshotMirrorEnableVolume(volume string) error {
 	return err
 }
 
+// cephFSSnapshotMirrorDisableVolume disables mirroring for a specific CephFS volume
 func cephFSSnapshotMirrorDisableVolume(volume string) error {
 	_, err := cephRun(cephFSSnapshotMirrorCmd([]string{
 		"disable", volume,
@@ -33,6 +36,7 @@ func cephFSSnapshotMirrorDisableVolume(volume string) error {
 	return err
 }
 
+// cephFSSnapshotMirrorAddPath enables mirroring for a specific path in a CephFS volume
 func cephFSSnapshotMirrorAddPath(ctx context.Context, volume string, path string) error {
 	_, err := cephRunContext(ctx, cephFSSnapshotMirrorCmd([]string{
 		"add", volume, path,
@@ -41,6 +45,7 @@ func cephFSSnapshotMirrorAddPath(ctx context.Context, volume string, path string
 	return err
 }
 
+// cephFSSnapshotMirrorRemovePath disables a path from being mirrored in a volume
 func cephFSSnapshotMirrorRemovePath(ctx context.Context, volume string, path string) error {
 	_, err := cephRunContext(ctx, cephFSSnapshotMirrorCmd([]string{
 		"remove", volume, path,
@@ -49,6 +54,7 @@ func cephFSSnapshotMirrorRemovePath(ctx context.Context, volume string, path str
 	return err
 }
 
+// cephFSSnapshotMirrorPeerCreate generates a bootstrap token for a cephfs mirroring peer
 func cephFSSnapshotMirrorPeerCreate(volume string, remoteName string, localName string) (string, error) {
 	output, err := cephRun(cephFSSnapshotMirrorCmd([]string{
 		"peer_bootstrap", "create", volume,
@@ -71,6 +77,7 @@ func cephFSSnapshotMirrorPeerCreate(volume string, remoteName string, localName 
 	return ret, nil
 }
 
+// cephFSSnapshotMirrorPeerImport imports a cephfs mirroring peer using the provided token
 func cephFSSnapshotMirrorPeerImport(volume string, token string) error {
 	_, err := cephRun(cephFSSnapshotMirrorCmd([]string{
 		"peer_bootstrap", "import", volume, token,
@@ -79,6 +86,7 @@ func cephFSSnapshotMirrorPeerImport(volume string, token string) error {
 	return err
 }
 
+// cephFSSnapshotMirrorPeerExists checks if a mirroring peer with the given remote name exists for the specified volume
 func cephFSSnapshotMirrorPeerExists(ctx context.Context, volume string, remoteName string) (bool, error) {
 	output, err := cephRunContext(ctx, cephFSSnapshotMirrorCmd([]string{
 		"peer_list", volume, "--format=json",
@@ -97,6 +105,8 @@ func cephFSSnapshotMirrorPeerExists(ctx context.Context, volume string, remoteNa
 		return false, err
 	}
 
+	logger.Debugf("CephFS snapshot peers found: %+v", peers)
+
 	for _, peer := range peers {
 		if peer.SiteName == remoteName {
 			return true, nil
@@ -106,12 +116,14 @@ func cephFSSnapshotMirrorPeerExists(ctx context.Context, volume string, remoteNa
 	return false, nil
 }
 
+// cephFSSnapshotMirrorList fetches the list of paths enabled for mirroring in a volume
 func cephFSSnapshotMirrorList(ctx context.Context, volume string) (string, error) {
 	return cephRunContext(ctx, cephFSSnapshotMirrorCmd([]string{
 		"ls", volume, "--format=json",
 	})...)
 }
 
+// cephFSSnapshotMirrorDaemonStatus fetches the cephfs mirroring daemon status
 func cephFSSnapshotMirrorDaemonStatus(ctx context.Context) (string, error) {
 	return cephRunContext(ctx, cephFSSnapshotMirrorCmd([]string{
 		"daemon", "status", "--format=json",
@@ -120,21 +132,11 @@ func cephFSSnapshotMirrorDaemonStatus(ctx context.Context) (string, error) {
 
 // Helper Methods
 
+// cephFSSnapshotMirrorCmd prefixes the snapshot mirror command
 func cephFSSnapshotMirrorCmd(args []string) []string {
 	cmd := []string{
 		"fs", "snapshot", "mirror",
 	}
 
 	return append(cmd, args...)
-}
-
-func dropEscapedChars(input string, dropStr string) string {
-	ret := fmt.Sprintf("%s", []byte(input))
-
-	logger.Debugf("Drop char Input (%s)", ret)
-
-	retval := strings.ReplaceAll(ret, dropStr, ``)
-
-	logger.Debugf("Drop char Output (%s)", retval)
-	return retval
 }
